linguistic evidentiality - initial implementation
The Naltok language is an artificial language, implemented by the ministries in their early days to facilitate scientific communication. Part of this is that it features grammaticalised evidentiality, which is conserved in translation. Any statement from a Naltok should also come with a clarifier explaining how they know what they're telling you.